Split difflib into pure-Python reference and public shim
Move the pure-Python implementation of difflib to Lib/_pydifflib.py and
turn Lib/difflib.py into a thin shim that re-exports its public API.
This mirrors the layout used by decimal/_pydecimal, datetime/_pydatetime,
and pickle/_pickle, where the public module dispatches to a faster C
implementation when available and the pure-Python module is preserved
as a self-contained reference for alternative Python implementations.

No public behaviour change.  ``Match`` is constructed with
``module='difflib'`` so its qualified name matches the public module.
